refactor: move the development file from src to outside
