+++ packages/core/index.html
@@ -2,12 +2,15 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="/src/demo/style.css">
+    <meta
+      name="viewport"
+      content="width=device-width, initial-scale=1.0"
+    />
+    <link rel="stylesheet" href="/development/style.css" />
     <title>Vite + TS</title>
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/demo/index.ts"></script>
+    <script type="module" src="/development/index.ts"></script>
   </body>
 </html>
