+++ packages/core/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../../tsconfig",
   "include": [
-    "src",
+    "src", "demo",
   ],
   "references": [
     {
