+++ packages/react/index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/demo/main.tsx"></script>
+    <script type="module" src="/development/main.tsx"></script>
   </body>
 </html>
