+++ packages/vue/index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/demo/main.ts"></script>
+    <script type="module" src="/development/main.ts"></script>
   </body>
 </html>
